Multithreaded step creation added and working... NOTE: Can only quit with ^c
Left debugging statements for you to see how it works

Uses pthread and has a two lists. One is populated while the other is running

On my machine this takes <20 frames so I start when there are 75 frames just in case

Sorry for all the globals. Hoipefully I can polish this up MUCH more. However, I think it is working quite well